Добавлено: - Установка samba(Author) - Фикс бага от Ruslandh с 302% установки пакетов.

+++ src/core.py
@@ -36,7 +36,7 @@ class Core:
                 os.system("clear")
 
             # Округляем процент до целого числа
-            percent = int(percent)
+
             print(process + ":")
 
             # Выводим прогресс бар длинной в 50 символов и заполненный на percent процентов
@@ -47,7 +47,7 @@ class Core:
                 else:
                     print("-", end="")
             # Выводим текущий процент выполнения
-            print("] - " + str(percent) + "%")
+            print("] - " + str(int(percent)) + "%")
             self.temp_percent = percent
             if percent == 100:
                 self.temp_percent = 0
@@ -116,15 +116,21 @@ class Core:
             return
 
         # Удаляем выбранные пакеты. Бюджет процентов на все пакеты 25%. С каждым удалённым пакетом увеличиваем
-        task_name = "Удаляю пакеты"
-        for i in range(0, len(remove_packages)):
-            try:
-                # Удаляем пакет
-                os.system("apt-get -qq remove" + remove_packages[i] + " &> /dev/null")
-                # Увеличиваем процент выполнения
-                self.progress_bar(task_name, self.temp_percent + (25 / len(remove_packages) * i))
-            except Exception:
-                log("Ошибка при удалении пакета " + remove_packages[i])
+        print("Собираюсь удалить пакеты: " + str(remove_packages))
+        print("Подтвердите удаление пакетов Да\нет")
+        answer = input().lower
+        if answer == "да" or answer == "д" or answer == "y" or answer == "":
+            os.system("clear")
+            task_name = "Удаляю пакеты"
+            print()
+            for i in range(0, len(remove_packages)):
+                try:
+                    # Удаляем пакет
+                    os.system("apt-get -qq remove" + remove_packages[i] + " &> /dev/null")
+                    # Увеличиваем процент выполнения
+                    self.progress_bar(task_name, self.temp_percent + (25 / len(remove_packages) * i))
+                except Exception:
+                    log("Ошибка при удалении пакета " + remove_packages[i])
 
         # Устанавливаем выбранные пакеты. Бюджет процентов на все пакеты 60%. С каждым установленным пакетом увеличиваем
         task_name = "Устанавливаю пакеты"
